requirements.txt
streamlit
paramiko
python-dotenv
pynput
pytest